chore(profile): Restructured routes
