+++ src/views/Profile/components/NftSelectionCard.tsx
