+++ src/views/Teams/components/NoProfileCard.tsx
@@ -22,7 +22,7 @@ const NoProfileCard = () => {
               {TranslateString(999, 'You can do this at any time by clicking on your profile picture in the menu')}
             </Text>
           </div>
-          <Button as={Link} to="/teams/profile" mt={['16px', null, 0]}>
+          <Button as={Link} to="/profile" mt={['16px', null, 0]}>
             {TranslateString(999, 'Set up now')}
           </Button>
         </Flex>
